requirements.txt
mss
pynput
pillow
opencv-python